Move constants into constants file
