+++ ui/src/kapacitor/constants/index.js
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 export const defaultRuleConfigs = {
   deadman: {
     period: '10m',
@@ -35,4 +37,25 @@ export const RULE_MESSAGE_TEMPLATES = {
   level: {label: "{{.Level}}", text: "Alert Level, one of: <code>INFO</code><code>WARNING</code><code>CRITICAL</code>"},
   fields: {label: `{{ index .Fields "value" }}`, text: "Map of fields. Use <code>&#123;&#123; index .Fields &quot;key&quot; &#125;&#125;</code> to get a specific field value"},
   time: {label: "{{.Time}}", text: "The time of the point that triggered the event"},
+};
+
+export const DEFAULT_ALERTS = ['http', 'tcp', 'exec', 'smtp', 'alerta'];
+
+export const DEFAULT_ALERT_PLACEHOLDERS = {
+  http: 'URL',
+  tcp: 'Address',
+  exec: 'Add command with arguments separated by spaces',
+  smtp: 'Add email addresses separated by spaces',
+  alerta: 'Paste Alerta tick script here',
+};
+
+export const ALERT_NODES_ACCESSORS = {
+  http: (rule) => _.get(rule, 'alertNodes[0].args[0]', ''),
+  tcp: (rule) => _.get(rule, 'alertNodes[0].args[0]', ''),
+  exec: (rule) => _.get(rule, 'alertNodes[0].args', []).join(' '),
+  smtp: (rule) => _.get(rule, 'alertNodes[0].args', []).join(' '),
+  alerta: (rule) => _.get(rule, 'alertNodes[0].properties', []).reduce((strs, item) => {
+    strs.push(`${item.name}('${item.args.join(' ')}')`);
+    return strs;
+  }, ['alerta()']).join('.'),
 };
